StudentPortal-master/new_requirements.txt
asgiref>>3.4.1
astroid>>2.2.5
autopep8>>1.5
build>>1.1.1
certifi>>2021.5.30
click>>8.1.7
colorama>>0.4.1
crispy-bootstrap5>>0.5
dj-database-url>>0.5.0
Django>>3.2.7
django-crispy-forms>>1.12.0
gunicorn>>20.1.0
isort>>4.3.21
lazy-object-proxy>>1.4.1
mccabe>>0.6.1
packaging>>24.0
Pillow>>8.3.2
pip-tools>>7.4.1
psycopg2-binary>>2.9.9
pycodestyle>>2.11.1
pyproject_hooks>>1.0.0
python-decouple>>3.5
pytz>>2023.3.post1
six>>1.16.0
sqlparse>>0.4.4
tomli>>2.0.1
typed-ast.>>1.5.5
typing-extensions>>3.10.0.2
validate_email>>1.3
wincertstore>>0.2
wrapt>>1.11.2
